CMP-1069
DIBS authTicket is modified to accept entire CARD object to support both TICKET and NEW CARD based operations.
In order to make sure that there is only one entry point in api/class/DIBS.php, it is necessary to ensure that the $ticket is wrapped in xml.
Dummy XML structure is used <callback><ticket>value</ticket></callback> as only ticket value is used for proecssing